==> 01_Analysen_Sammelboegen.Rmd <==
---
title: "Ergebnisse aus den U-Sammelbögen"
output: 
  pdf_document:
    latex_engine: lualatex # Per default wird pdflatex verwendet. fontspec braucht allerdings lualatex.
header-includes:
  - \usepackage[ngerman]{babel} # Sprache und Format auf Deutsch. Nicht besonders wichtig für uns aber schadet auch nicht.
  - \usepackage[T1]{fontenc} # Für Umlaute
  - \usepackage[utf8]{inputenc} # Alles UTF-8
  - \usepackage{fontspec} # Für Calibri
  - \setmainfont{Calibri}
  - \usepackage{float}
geometry: margin = 2cm
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
#tinytex::install_tinytex() #notwendig, wenn keine LateX-Distro oder für Update

library(tidyverse)
library(janitor)

library(psych)
library(readxl)
library(tinytex) # Notwendig damit LaTeX-Packages richtig installiert werden

library(knitr)
library(kableExtra)
library(table1)
```

```{r data, include=FALSE}
## Datensätze
setwd('O:/U4951/Consulimus_Abschlussdatenlieferung/03_Sammelbögen/01_Data')
 
u2 <- read.csv('u2.csv')
u3 <- read.csv('u3.csv')
u4 <- read.csv('u4.csv')
u5 <- read.csv('u5.csv')
u6 <- read.csv('u6.csv')
u7 <- read.csv('u7.csv')
u7a <- read.csv('u7a.csv')
u8 <- read.csv('u8.csv')
u9 <- read.csv('u9.csv')
```


```{r functions, include=FALSE}
#### ja /nein ####
rownames_janein <- c('ja', 'nein', 'valides n', 'fehlend', 'gesamt n')

tbl_janein <-function(x) {
  # Verarbeitet Variable mit zwei Ausprägungen und gibt Tabelle mit absoluten und 
  # relativen Zahlen aus
  x[x==0] <- NA
  tbl <- janitor::tabyl(x, show_na = TRUE) 
  
  # change name if no NAs and valid_percent is not created
  if (!('valid_percent' %in% names(tbl))) {
    tbl <- tbl %>% mutate(valid_percent=percent)
  }
    tbl <- tbl %>% 
    mutate(freq=valid_percent*100) %>% 
    select(-percent, -valid_percent)
  
  # if no NAs: set n of NAs to zero
   if (is.na(tbl[3,2])) {
    tbl[3,2] <- 0
  }  
  # valids n, fehlend, gesamt hinzufuegen
  tbl[4,] <- tbl[3,] #NAs verschieben
  tbl[3,2:3] <- colSums(tbl[1:2,2:3]) #valides n
  tbl[5,2] <- sum(tbl[3:4,2]) #gesamt n
  return(tbl)
}

ges_janein <- function(df){
  # Funktion erstellt gesamt-Spalten [n] & [%]
  df <- df %>%  mutate(n_ges = rowSums(across(contains('n'))))
  a <- df$n_ges[1] / df$n_ges[3]*100
  b <- df$n_ges[2] / df$n_ges[3]*100
  c <- sum(a+b)
  df$freq_ges <- as.numeric(c(a,b,c, NA, NA))
  df <- df %>% 
  mutate_at(vars(contains('freq')), ~ sprintf("%.1f", round(., 1))) %>% 
  mutate_at(vars(contains('freq')), ~ gsub("\\.", ",", .)) %>% 
  mutate_all(na_if,'NA')
  return(df)
}

#### mehr Ausprägungen ####
tbl_mehrf <-function(x) {
  # Verarbeitet Variable mit mehreren Ausprägungen und gibt Tabelle mit absoluten und 
  # relativen Zahlen aus
  # x[x==0] <- NA
  tbl1 <- as.data.frame(table(x, useNA='always')) 
  tbl2 <- as.data.frame(round(prop.table(table(x)),6)*100)
  #tbl2$Freq <- paste(tbl2$Freq, '%') #doch kein %-Zeichen
  tbl2 <- rbind(tbl2, NA)
  tbl1 <- cbind(tbl1, tbl2[2])
  # zusätzliche Zeilen einfügen
  tbl1[nrow(tbl1)+1,] <- tbl1[nrow(tbl1),] #NAs verschieben
  tbl1[nrow(tbl1)-1,2:3] <- colSums(tbl1[1:(nrow(tbl1)-2),2:3]) #valides n
  #tbl1[nrow(tbl1)-1,2:3] <- 100.0 #valides n
  tbl1[nrow(tbl1)+1,2] <- sum(tbl1[(nrow(tbl1)-1):(nrow(tbl1)),2]) #gesamt n
  tbl1
}

ges_mehrf <- function(df){
  # Funktion erstellt gesamt-Spalten [n] & [%] für mehrere Auspärgungen
  df <- df %>% mutate(`Ges. [n]` = rowSums(across(contains('[n]'))))
  df$`Ges. [%]` <- NA
  df$`Ges. [%]`[1:(nrow(df)-3)] <- df$`Ges. [n]`[1:(nrow(df)-3)] / df$`Ges. [n]`[(nrow(df)-2)]*100
  # df$`Ges. [%]`[(nrow(df)-2)] <- sum(df$`Ges. [%]`[1:(nrow(df)-3)])
  df$`Ges. [%]`[(nrow(df)-2)] <- 100.0
  df
}

#### Sonstige ####
# Wrapper-Funktion, um die decimal marks nicht jedes mal separat anzugeben 
kbl2 = function(...) {
  kableExtra::kbl(..., format.args = list(decimal.mark = ',', big.mark = "."), linesep = "")
}

# NAs werden nicht in den Tabellen dargestellt
options(knitr.kable.NA = '')

# erstelle String für Tebellen-alignment (alle ab zweiter Splate rechtsbündig)
align_string <- function(df) {
  paste('l', paste(rep('r', (length(df)-1)), collapse=''), sep='')
}

# kontinuierliche Variable
tbl_kont <-function(x) {
  tbl1 <- describe(x)
  tbl1 <- select(tbl1, n, mean, sd, median, min, max)
  names(tbl1) <- c('n', 'M', 'SD', 'Md', 'Min', 'Max')
  tbl1
}
```

# 3. Rücklauf der Erhebungen und Stichprobencharakterisierung

## 3.1 Erhebungen in den Arztpraxen

### 3.1.1 Arztfragebogen und U-Sammelbögen

Anzahl untersuchter Kinder nach U-Untersuchung
```{r n, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- u2 %>% select(id, U)
tbl2 <- u3 %>% select(id, U)
tbl3 <- u4 %>% select(id, U)
tbl4 <- u5 %>% select(id, U)
tbl5 <- u6 %>% select(id, U)
tbl6 <- u7 %>% select(id, U)
tbl7 <- u7a %>% select(id, U)
tbl8 <- u8 %>% select(id, U)
tbl9 <- u9 %>% select(id, U)
tbl <- bind_rows(tbl1, tbl2, tbl3, tbl4, tbl5, tbl6, tbl7, tbl8, tbl9)

tbl <- tbl %>% tabyl(U) %>% mutate(percent=percent*100) %>% adorn_totals('row', name='Ges.')

tbl %>%
  kbl2(booktabs = T, centering=F, digits=1,
       col.names=c("U-Untersuchung", "n", "%")) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(9, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

```


```{r eval=FALSE, include=FALSE}
tbl <- bind_rows(tbl1, tbl2, tbl3, tbl4, tbl5, tbl6, tbl7, tbl8, tbl9)

tbl <- tbl %>% 
  group_by(id) %>% 
  summarize(n=n())

psych::describe(tbl$n)
```


## 3.2 Erhebungen bei den Eltern

### 3.2.1 Elternfragebogen

Studienunterlagen an Eltern ausgeteilt - Kreuz gesetzt 

```{r echo=FALSE, message=FALSE, warning=FALSE}
u2$F14 <- ifelse(u2$F14==0, 99, u2$F14)
tbl1 <- tbl_janein(u2$F14)
u3$F14 <- ifelse(u3$F14==0, 99, u3$F14)
tbl2 <- tbl_janein(u3$F14)
u4$F12 <- ifelse(u4$F12==0, 99, u4$F12)
tbl3 <- tbl_janein(u4$F12)
u5$F21 <- ifelse(u5$F21==0, 99, u5$F21)
tbl4 <- tbl_janein(u5$F21)
u6$F23 <- ifelse(u6$F23==0, 99, u6$F23)
tbl5 <- tbl_janein(u6$F23)
u7$F21 <- ifelse(u7$F21==0, 99, u7$F21)
tbl6 <- tbl_janein(u7$F21)
u7a$F30 <- ifelse(u7a$F30==0, 99, u7a$F30)
tbl7 <- tbl_janein(u7a$F30)
u8$F41 <- ifelse(u8$F41==0, 99, u8$F41)
tbl8 <- tbl_janein(u8$F41)
u9$F30 <- ifelse(u9$F30==0, 99, u9$F30)
tbl9 <- tbl_janein(u9$F30)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c("ja", "nein", "valides n")) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


# 4. Ergebnisse der Evaluation von Abschnitt B (Früherkennungsuntersuchungen) der Kinder-Richtlinie 

## 4.1	Untersuchung der Augen (U2-U9)

### 4.1.2 Inspektion und Prüfung im durchfallenden Licht (U2 und U3)

#### 4.1.2.1 Häufigkeit der Durchführung 

 \newline
Inspektion (auf morphologische Auffälligkeiten) durchgeführt? 
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F9)
tbl2 <- tbl_janein(u3$F9)
tbl <- bind_cols(tbl1, tbl2[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 3))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U2'=2, 'U3'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


Prüfung im durchfallenden Licht durchgeführt?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F11)
tbl2 <- tbl_janein(u3$F11)
tbl <- bind_cols(tbl1, tbl2[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 3))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U2'=2, 'U3'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

#### 4.1.2.2	Häufigkeit der Feststellung von Auffälligkeiten

 \newline
Morphologische Auffälligkeiten festgestellt? (Feststellung durch Inspektion). Ausgewertet nur dann, wenn eine Inspektion durchgeführt wurde <br>
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F10[u2$F9==1])
tbl2 <- tbl_janein(u3$F10[u3$F9==1])
tbl <- bind_cols(tbl1, tbl2[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 3))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U2'=2, 'U3'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

\pagebreak

Transilluminationsauffälligkeiten festgestellt? (Feststellung durch Prüfung im durchfallenden Licht)
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F12[u2$F11==1])
tbl2 <- tbl_janein(u3$F12[u3$F11==1])
tbl <- bind_cols(tbl1, tbl2[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 3))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U2'=2, 'U3'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


##### 4.1.2.3	Häufigkeit von Überweisungen

 \newline
Überweisung zum Augenarzt? Alle ausgewertet
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F13)
tbl2 <- tbl_janein(u3$F13)
tbl <- bind_cols(tbl1, tbl2[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 3))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U2'=2, 'U3'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Überweisung zum Augenarzt? Nur ausgewertet wenn Inspektion oder Prüfung im durchfallenden Licht durchgeführt und entweder in der Inspektion oder in der Prüfung im durchfallenden Licht Auffälligkeiten festgestellt wurden. 
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F13[(u2$F9 == 1 &  u2$F10 == 1) | (u2$F11 == 1 &  u2$F12 == 1)])
tbl2 <- tbl_janein(u3$F13[(u3$F9 == 1 &  u3$F10 == 1) | (u3$F11 == 1 &  u3$F12 == 1)])
tbl <- bind_cols(tbl1, tbl2[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 3))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U2'=2, 'U3'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

### 4.1.3	Brückner-Test (U4–U7)

#### 4.1.3.1	Häufigkeit der Durchführung

 \newline
Brückner-Test durchgeführt?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u4$F9)
tbl2 <- tbl_janein(u5$F9)
tbl3 <- tbl_janein(u6$F9)
tbl4 <- tbl_janein(u7$F9)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 5))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U4'=2, 'U5'=2, 'U6'=2, 'U7'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Davon: Prüfung auf Transilluminationsunterschied mit einem binokularen Autorefraktometer durchgeführt? Nur ausgewertet, wenn ein Brückner-Test durchgeführt wurde
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u4$F10[u4$F9==1])
tbl2 <- tbl_janein(u5$F10[u5$F9==1])
tbl3 <- tbl_janein(u6$F10[u6$F9==1])
tbl4 <- tbl_janein(u7$F10[u7$F9==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 5))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U4'=2, 'U5'=2, 'U6'=2, 'U7'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```



#### 4.1.3.2	Häufigkeit der Feststellung von Auffälligkeiten

 \newline
Transilluminationsunterschied festgestellt? Nur ausgewertet, wenn ein Brückner-Test durchgeführt wurde.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u4$F11[u4$F9==1])
tbl2 <- tbl_janein(u5$F11[u5$F9==1])
tbl3 <- tbl_janein(u6$F11[u6$F9==1])
tbl4 <- tbl_janein(u7$F11[u7$F9==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein

a <- tbl %>% filter(x %in% c('ja', 'nein')) %>% mutate(y='alle U4-U7 Untersuchungen') #für Abbildung

names(tbl) <- c('Antwort', rep(c('n', '%'), 5)) 

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U4'=2, 'U5'=2, 'U6'=2, 'U7'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

Transilluminationsunterschied festgestellt? Nur ausgewertet, wenn ein Brückner-Test *mit* einem binokularen Autorefraktometer durchgeführt wurde.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u4$F11[u4$F9==1 & u4$F10==1])
tbl2 <- tbl_janein(u5$F11[u5$F9==1 & u5$F10==1])
tbl3 <- tbl_janein(u6$F11[u6$F9==1 & u6$F10==1])
tbl4 <- tbl_janein(u7$F11[u7$F9==1 & u7$F10==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein

b <- tbl %>% filter(x %in% c('ja', 'nein')) %>% mutate(y='mit Autorefraktometer') #für Abbildung

names(tbl) <- c('Antwort', rep(c('n', '%'), 5))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U4'=2, 'U5'=2, 'U6'=2, 'U7'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

Transilluminationsunterschied festgestellt? Nur ausgewertet, wenn ein Brückner-Test *ohne* binokulares Autorefraktometer durchgeführt wurde.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u4$F11[u4$F9==1 & u4$F10==2])
tbl2 <- tbl_janein(u5$F11[u5$F9==1 & u5$F10==2])
tbl3 <- tbl_janein(u6$F11[u6$F9==1 & u6$F10==2])
tbl4 <- tbl_janein(u7$F11[u7$F9==1 & u7$F10==2])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein

c <- tbl %>% filter(x %in% c('ja', 'nein')) %>% mutate(y='ohne Autorefraktometer') #für Abbildung

names(tbl) <- c('Antwort', rep(c('n', '%'), 5))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U4'=2, 'U5'=2, 'U6'=2, 'U7'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


```{r plot1, echo=FALSE, include=FALSE, message=FALSE, warning=FALSE}
### GRAFIK TRANSILLUMINATIONSUNTERSCHEIDE

tmp <- bind_rows(a, b, c)
tmp <- tmp %>% select(x, y, contains('n'))
names(tmp) <- c('x', 'y', 'U4', 'U5', 'U6', 'U7', 'ges.')

tmp <- tmp %>% 
  pivot_longer(U4:ges., names_to='U', values_to='n') %>% 
  group_by(U, y) %>% 
  mutate(perc = n / sum(n)) %>% 
  ungroup()

p <- tmp %>% 
  filter(x == 'ja') %>%
  ggplot(aes(y=perc, 
             x=factor(y, levels=c('alle U4-U7 Untersuchungen', 'ohne Autorefraktometer', 'mit Autorefraktometer')), 
             fill=factor(y, levels=c('alle U4-U7 Untersuchungen', 'ohne Autorefraktometer', 'mit Autorefraktometer')), 
             label = scales::percent(perc, accuracy=.1, decimal.mark = ","))) + 
  geom_bar(stat="identity") +
  scale_y_continuous(limits = c(0.0, 0.2), labels = scales::percent_format(accuracy=.1, decimal.mark = ",")) +
  geom_text(position = position_dodge(width = .9),vjust = -0.5, size = 4.5, color="#00386a") + 
  scale_fill_manual(values=c('#00386a', '#eda63b', '#BED600')) +
  facet_wrap(~ factor(U, levels=c('U4', 'U5', 'U6', 'U7', 'ges.')), nrow=1) +
  ylab('Anteil Untersuchungen') +
  theme_bw() +
  theme(strip.background = element_rect(fill="#cbd3b2"),
        axis.title.x = element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x = element_blank(),
        axis.title.y = element_text(color="#00386a", family = "Verdana", size = 14),
        axis.text.y = element_text(color="#00386a", family = "Verdana", size = 14),
        strip.text.x = element_text(color="#00386a", family = "Verdana", size = 14),
        legend.position="bottom",
        legend.title = element_blank(), 
        legend.text = element_text(color="#00386a", family = "Verdana", size = 14),
        plot.title = element_blank())

bmp(file="04_Abbildungen/Transilluminationsunterschiede.bmp", width = 750, height = 500, type="windows")
print(p)
dev.off()

```


\pagebreak

### 4.1.4	Inspektion, Pupillenstatus, Hornhautreflexbildchen, Stereo-Test und Formenwiedererkennungstest (U7a–U9)

#### 4.1.4.1	Häufigkeit der Durchführung 

 \newline
Inspektion durchgeführt?
```{r Inspektion, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F9)
tbl2 <- tbl_janein(u8$F9)
tbl3 <- tbl_janein(u9$F9)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Pupillenstatus geprüft?
```{r Pupillenstatus, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F11)
tbl2 <- tbl_janein(u8$F11)
tbl3 <- tbl_janein(u9$F11)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Hornhautreflexbildchen durchgeführt?
```{r Hornhautreflexbildchen, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F13)
tbl2 <- tbl_janein(u8$F13)
tbl3 <- tbl_janein(u9$F13)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Test mit einem  Autorefraktometer durchgeführt? (z. B. von Plusoptix, Welch Allyn o. a. Herstellern). Nur ausgewertet, wenn Hornhautreflexbildchen durchgeführt. 
```{r Hornhautreflexbildchen_plusoptix, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F14[u7a$F13==1])
tbl2 <- tbl_janein(u8$F14[u8$F13==1])
tbl3 <- tbl_janein(u9$F14[u9$F13==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Stereo-Test durchgeführt?
```{r Stereo_Test, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F16)
tbl2 <- tbl_janein(u8$F16)
tbl3 <- tbl_janein(u9$F16)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Test mit einem binokularem Autorefraktometer durchgeführt? (z. B. von Plusoptix, Welch Allyn o.a. Herstellern)
```{r Stereo_Test_plusoptix, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F17[u7a$F16==1])
tbl2 <- tbl_janein(u8$F17[u8$F16==1])
tbl3 <- tbl_janein(u9$F17[u9$F16==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Sehtest als monokulare Prüfung mittels nonverbalem Formenwiedererkennungstest durchgeführt?
```{r nonverbForm, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F19)
tbl2 <- tbl_janein(u8$F19)
tbl3 <- tbl_janein(u9$F19)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


  \linebreak

Anteil der Kinder in der U7a, U8 bzw. U9, bei denen keiner, einer, zwei, drei, etc. der Sehtests durchgeführt wurden.
```{r message=FALSE, warning=FALSE}
augen_recode <- function(x) {
  case_when(
    x == 1 ~ 1, 
    x == 2 ~ 0,
    x == 0 ~ as.numeric(NA)
  )
}

U7a_tmp <- u7a %>% 
  mutate_at(vars(F9, F11, F13, F16, F19), augen_recode) %>% 
  select(X, F9, F11, F13, F16, F19) %>% 
  group_by(X) %>%
  mutate(sum_test_u7a=sum(F9 + F11 + F13 + F16 + F19, na.rm=TRUE),
         sum_test_u7a=ifelse(is.na(F9)|is.na(F11)|is.na(F13)|is.na(F16)|is.na(F19), 
                             NA, sum_test_u7a)) %>% 
  ungroup() %>%
  count(sum_test_u7a)
  
U8_tmp <- u8 %>% 
  mutate_at(vars(F9, F11, F13, F16, F19), augen_recode) %>% 
  select(X, F9, F11, F13, F16, F19) %>% 
  group_by(X) %>%
  mutate(sum_test_u8=sum(F9 + F11 + F13 + F16 + F19, na.rm=TRUE),
         sum_test_u8=ifelse(is.na(F9)|is.na(F11)|is.na(F13)|is.na(F16)|is.na(F19), 
                             NA, sum_test_u8)) %>%   
  ungroup() %>%
  count(sum_test_u8)

U9_tmp <- u9 %>% 
  mutate_at(vars(F9, F11, F13, F16, F19), augen_recode) %>% 
  select(X, F9, F11, F13, F16, F19) %>% 
  group_by(X) %>%
  mutate(sum_test_u9=sum(F9 + F11 + F13 + F16 + F19, na.rm=TRUE),
         sum_test_u9=ifelse(is.na(F9)|is.na(F11)|is.na(F13)|is.na(F16)|is.na(F19), 
                             NA, sum_test_u9)) %>%   
  ungroup() %>%
  count(sum_test_u9)

tbl <- bind_cols(U7a_tmp, U8_tmp[2], U9_tmp[2]) 
names(tbl) <- c('tests', 'u7a', 'u8', 'u9')

# fehlende erstmal extrahieren
tmp <- tbl %>% filter(is.na(tests))
tbl <- tbl %>% filter(!is.na(tests))

tbl <- tbl %>%
  group_by(tests)%>% 
  mutate(ges = sum(u7a, u8, u9)) %>% 
  ungroup() %>% 
  mutate(u7a_perc = u7a / sum(u7a) * 100, 
         u8_perc = u8 / sum(u8) * 100, 
         u9_perc = u9 / sum(u9) * 100, 
         ges_perc = ges / sum(ges) * 100) %>% 
  select(tests, contains('u7a'), contains('u8'), contains('u9'), contains('ges')) %>% 
  janitor::adorn_totals('row', name='valides n')


# fehlende Werte manuell ergänzen
tbl[nrow(tbl) + 1,] <- 
  c('fehlend', tmp[2], NA, tmp[3], NA, tmp[4], NA, sum(tmp[2]+ tmp[3]+ tmp[4]), NA)

tbl[nrow(tbl) + 1,] <- 
  list('gesamt n', 5290, NA, 5349, NA, 5002, NA, 15641, NA)

names(tbl) <- c('Anzahl Sehtests', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl), digits=1) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(6, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Anteil der Kinder in der U7a, U8 bzw. U9, bei denen die Hornhautreflexbildchen *nicht* geprüft wurden, aber der Stereo-Test oder der Formenwiedererkennungstest durchgeführt wurden. Nur ausgewertet, wenn Hornhautreflexbildchen *nicht* geprüft wurden.
```{r message=FALSE, warning=FALSE}
u7a <- u7a %>% mutate(test = ifelse(F16 == 1 | F19 == 1, '1', '2'), 
                      test = ifelse(F16 == 0 | F19 == 0, NA, test))
u8 <- u8 %>% mutate(test = ifelse(F16 == 1 | F19 == 1, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))
u9 <- u9 %>% mutate(test = ifelse(F16 == 1 | F19 == 1, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))


tbl1 <- tbl_janein(u7a$test[u7a$F13 == 2])
tbl2 <- tbl_janein(u8$test[u8$F13 == 2])
tbl3 <- tbl_janein(u9$test[u9$F13 == 2])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


  \linebreak

Anteil der Kinder in der U7a, U8 bzw. U9, bei denen Stereo-Test UND Formenwiedererkennungstest durchgeführt wurden, NUR der Stereo-Test, NUR der Formenwiedererkennungstest oder KEINER von beiden Tests durchegführt wurde
```{r message=FALSE, warning=FALSE}
u7a <- u7a %>% mutate(test = ifelse(F16 == 1 & F19 == 1, '1', '2'), 
                      test = ifelse(F16 == 0 | F19 == 0, NA, test))
u8 <- u8 %>% mutate(test = ifelse(F16 == 1 & F19 == 1, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))
u9 <- u9 %>% mutate(test = ifelse(F16 == 1 & F19 == 1, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))

tbl1 <- tbl_janein(u7a$test)
tbl2 <- tbl_janein(u8$test)
tbl3 <- tbl_janein(u9$test)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tmp1 <- tbl

#***************************
u7a <- u7a %>% mutate(test = ifelse(F16 == 1 & F19 == 2, '1', '2'), 
                      test = ifelse(F16 == 0 | F19 == 0, NA, test))
u8 <- u8 %>% mutate(test = ifelse(F16 == 1 & F19 == 2, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))
u9 <- u9 %>% mutate(test = ifelse(F16 == 1 & F19 == 2, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))

tbl1 <- tbl_janein(u7a$test)
tbl2 <- tbl_janein(u8$test)
tbl3 <- tbl_janein(u9$test)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tmp2 <- tbl

#*****************
u7a <- u7a %>% mutate(test = ifelse(F16 == 2 & F19 == 1, '1', '2'), 
                      test = ifelse(F16 == 0 | F19 == 0, NA, test))
u8 <- u8 %>% mutate(test = ifelse(F16 == 2 & F19 == 1, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))
u9 <- u9 %>% mutate(test = ifelse(F16 == 2 & F19 == 1, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))

tbl1 <- tbl_janein(u7a$test)
tbl2 <- tbl_janein(u8$test)
tbl3 <- tbl_janein(u9$test)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tmp3 <- tbl


#*******************
u7a <- u7a %>% mutate(test = ifelse(F16 == 2 & F19 == 2, '1', '2'), 
                      test = ifelse(F16 == 0 | F19 == 0, NA, test))
u8 <- u8 %>% mutate(test = ifelse(F16 == 2 & F19 == 2, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))
u9 <- u9 %>% mutate(test = ifelse(F16 == 2 & F19 == 2, '1', '2'), 
                    test = ifelse(F16 == 0 | F19 == 0, NA, test))

tbl1 <- tbl_janein(u7a$test)
tbl2 <- tbl_janein(u8$test)
tbl3 <- tbl_janein(u9$test)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tmp4 <- tbl

tbl <- bind_rows(tmp1[1,], tmp2[1,], tmp3[1,], tmp4[1,], tmp1[3:5,])
tbl$Antwort <- c('beide Tests', 'Stereo Test', 'Formenwiedererkennungstest', 'keiner von beiden Tests', 'valides n', 'fehlend', 'gesamt n')

names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(4, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```







#### 4.1.4.2	Häufigkeit der Feststellung von Auffälligkeiten 

 \newline
morphologische Auffälligkeit oder Nystagmus oder Kopffehlhaltung? Nur ausgewertet, wenn Inspektion durchgeführt. 
```{r morphAuff, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F10[u7a$F9==1])
tbl2 <- tbl_janein(u8$F10[u8$F9==1])
tbl3 <- tbl_janein(u9$F10[u9$F9==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

\pagebreak

Größe, Form, Lichtreaktion der Pupillen auffällig? Nur ausgewertet, wenn Pupillenstatus geprüft.
```{r Pupillen, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F12[u7a$F11==1])
tbl2 <- tbl_janein(u8$F12[u8$F11==1])
tbl3 <- tbl_janein(u9$F12[u9$F11==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Strabismus festgestellt? Nur ausgeweret, wenn Hornhautreflexbildchen durchgeführt.
```{r Strabismus, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F15[u7a$F13==1])
tbl2 <- tbl_janein(u8$F15[u8$F13==1])
tbl3 <- tbl_janein(u9$F15[u9$F13==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak
  
Strabismus festgestellt? Nur ausgeweret, wenn Hornhautreflexbildchen *mit* binokularem Autorefraktometer durchgeführt.
```{r Strabismus_plusoptix, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F15[u7a$F13==1 & u7a$F14==1])
tbl2 <- tbl_janein(u8$F15[u8$F13==1 & u8$F14==1])
tbl3 <- tbl_janein(u9$F15[u9$F13==1 & u9$F14==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Strabismus festgestellt? Nur ausgeweret, wenn Hornhautreflexbildchen *ohne* binokulares Autorefraktometer durchgeführt.
```{r Strabismus_ohneplusoptix, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F15[u7a$F13==1 & u7a$F14==2])
tbl2 <- tbl_janein(u8$F15[u8$F13==1 & u8$F14==2])
tbl3 <- tbl_janein(u9$F15[u9$F13==1 & u9$F14==2])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


\pagebreak

Stereo-Test auffällig? Nur ausgewertet, wenn Stereo-Test durchegführt.
```{r Stereo_Test_Ergebnis, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F18[u7a$F16==1])
tbl2 <- tbl_janein(u8$F18[u8$F16==1])
tbl3 <- tbl_janein(u9$F18[u9$F16==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Stereo-Test auffällig? Nur ausgewertet, wenn Stereo-Test *mit* binokularem Autorefraktometer durchegführt.
```{r Stereo_Test_Ergebnis_plusoptix, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F18[u7a$F16==1 & u7a$F17==1])
tbl2 <- tbl_janein(u8$F18[u8$F16==1 & u8$F17==1])
tbl3 <- tbl_janein(u9$F18[u9$F16==1 & u9$F17==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Stereo-Test auffällig? Nur ausgewertet, wenn Stereo-Test *ohne* binokularem Autorefraktometer durchegführt.
```{r Stereo_Test_Ergebnis_ohneplusoptix, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F18[u7a$F16==1 & u7a$F17==2])
tbl2 <- tbl_janein(u8$F18[u8$F16==1 & u8$F17==2])
tbl3 <- tbl_janein(u9$F18[u9$F16==1 & u9$F17==2])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Sehschwäche o. Rechts-Links-Differenz festgestellt. Nur ausgewertet, wenn ein Sehtest mittels nonverbalem Formenwiedererkennungstest durchgeführt wurde.
```{r Sehschwaeche, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u7a$F20[u7a$F19==1])
tbl2 <- tbl_janein(u8$F20[u8$F19==1])
tbl3 <- tbl_janein(u9$F20[u9$F19==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 4))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U7a'=2, 'U8'=2, 'U9'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


## 4.2	Untersuchung des Hörvermögens (U8)

### 4.2.2	Hörtest mittels Screeningaudiometrie (U8)

#### 4.2.2.1	Häufigkeit der Durchführung 

 \newline
Prüfung mittels Tonschwellenaudiometrie durchgeführt?
```{r Audiometrie, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u8$F21)
tbl <- ges_janein(tbl)

tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U8'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```
<br>

#### 4.2.2.2	Gründe keine Audiometrie durchzuführen

 \newline
Falls Prüfung mittels Tonschwellenaudiometrie NICHT durchgeführt: Warum nicht? (Mehrfachnennungen möglich). Nur ausgewertet, wenn keine Prüfung mittels Tonschwellenaudiometrie durchgeführt.

```{r Audiometrie_Gruende, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- u8 %>% 
  filter(F21==2) %>% 
  select(F25:F31) %>% 
  mutate(across(everything(), ~replace(., . == 99 , 0))) %>% 
  mutate(valid=ifelse(rowSums(. ,na.rm=TRUE )>=1,1,0), 
         missing=ifelse(rowSums(., na.rm=TRUE)==0,1,0), 
         ges=1) %>%
  summarise_all(sum, na.rm = TRUE) %>%
  set_names(linebreak(c(
    "01 Kind bereits in HNO-ärztlicher und/oder pädaudiologischer Behandlung",
    "02 Sprachbarriere",
    "03 keine Sprachbarriere, aber Kind mit den Anweisungen im Test überfordert",
    "04 Kind nicht überfordert, kooperiert aber nicht",
    "05 fehlende Kooperation der Eltern",
    "06 klinische Gründe (bspw. otitis media)",
    "07 sonstiger Grund",
    "97 valides n", 
    "98 fehlend",
    "99 gesamt n"))) %>% 
  rownames_to_column() %>% 
  pivot_longer(-rowname, names_to = "var", values_to= "n") %>%
  #spread(Gruppe_, n) # nur wenn mehrere Gruppen
  select(var, n) %>% 
  mutate(var=str_sub(var, 4L)) %>%
  set_names(c("Grund", "n"))

# der Häufigkeit nach sortieren
tbl1 <- tbl[1:6,]
tbl1 <- tbl1 %>% arrange(desc(n))
tbl[1:6,] <- tbl1

# Prozentangaben berechnen
tbl$freq <- (tbl$n / tbl$n[tbl$Grund=="valides n"])*100
# ungültige löschen
tbl[nrow(tbl),'freq'] <- NA
tbl[nrow(tbl)-1,'freq'] <- NA
tbl[nrow(tbl)-2,'freq'] <- NA

tbl %>% 
  select('Grund', 'n') %>%
  slice_tail(n=3) %>% 
    kbl2(booktabs = T, escape = F, digits=1, 
       col.names=c("Fallzahl", "n")) %>%
    kable_styling(latex_options = "HOLD_position", position = "left") 

tbl %>% 
  slice_head(n=7) %>%
  kbl2(booktabs = T, escape = F, digits=1, 
       col.names=c("Grund", "n", "\\% U-Untersuchungen")) %>%
    kable_styling(latex_options = "HOLD_position", position = "left")
```

\pagebreak

Falls Prüfung mittels Tonschwellenaudiometrie NICHT durchgeführt: Sonstige Gründe gegen die Durchführung (Freitextantworten, als Nenner wird einerseits die Anzahl von Kindern, bei denen mind. ein sonstiger Grund genannt wurde (n=219) und andererseist die Gesamtzahl von validen Antworten auf die Frage nach Gründen gegen die Durchführung einer Tonschwellenaudiometrie verwendet).

```{r Audiometrie_Gruende_Freitexte, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- read_xlsx('O:/U4951/Consulimus_Abschlussdatenlieferung/03_Sammelbögen/02_Freitexte/Freitexte_GruendeAudiometrie_CodiersystemMAXQDA_grb.xlsx', sheet='R_input')

tbl <- tbl %>% mutate(rownr=as.character(row_number())) %>% 
  arrange(match(rownr, c('1', '3', '2', '4', '5', '7', '9', '8', '6', '10', '11', '12'))) %>% 
  select(-rownr) %>% 
  mutate(freq1=n/219*100,
         freq2=n/903*100)

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl), digits=1, 
             col.names=c("Grund", "n", "% Sonstige", "% U-Untersuchungen")) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(c(5,10), extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


#### 4.2.2.3	Häufigkeit der Feststellung von Auffälligkeiten

 \newline
Tonschwellenaudiometrie mit auffälligem Befund? Nur ausgewertet, wenn eine Prüfung mittels Tonschwellenaudiometrie durchgeführt wurde. 
```{r Auffaelligkeiten_Audiometrie, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u8$F22[u8$F21==1])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U8'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


Bei auffälligem Befund: Gab es vor der heutigen Audiometrie bereits Anhaltspunkte für eine Störung des Hörvermögens? Nur ausgewertet, wenn eine Tonschwellenaudiometrie durchgeführt und ein auffälliger Befund angegeben wurde. 
```{r Auffaelligkeiten_Audiometrie_erstmalig, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u8$F24[u8$F21==1 & u8$F22==1])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U8'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


\pagebreak
#### 4.2.2.4	Häufigkeit von Überweisungen

 \newline
Bei auffälligem Befund: Überweisung an HNO-Arzt oder Pädaudiologen? Nur ausgewertet, wenn eine Tonschwellenaudiometrie durchgeführt und ein auffälliger Befund angegeben wurde.
```{r Ueberweisungen_Audiometrie, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u8$F23[u8$F21==1 & u8$F22==1])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U8'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


  \linebreak

Bei auffälligem Befund: Überweisung an HNO-Arzt oder Pädaudiologen? Nur ausgewertet, wenn eine Tonschwellenaudiometrie durchgeführt und ein **erstmaliger** auffälliger Befund angegeben wurde.
```{r Ueberweisungen_Audiometrie_erstmalig, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u8$F23[u8$F21==1 & u8$F22==1 & u8$F24==2])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U8'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


Bei auffälligem Befund: Überweisung an HNO-Arzt oder Pädaudiologen? Nur ausgewertet, wenn eine Tonschwellenaudiometrie durchgeführt und **bereits vorher** Anhaltspunkte für ein Störung des Hörvermögens bestanden.
```{r Ueberweisungen_Audiometrie_wiederholt, echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u8$F23[u8$F21==1 & u8$F22==1 & u8$F24==1])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U8'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

\pagebreak

## 4.3	Untersuchung der Zähne und der Schleimhaut (U6)

### 4.3.2	Auffälligkeiten an Zähnen und Schleimhaut (U6)

#### 4.3.2.1	Häufigkeit der Feststellung von Auffälligkeiten

 \newline
Auffälligkeiten an Zähnen oder der Schleimhaut festgestellt?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u6$F12)

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U6'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

#### 4.3.2.2	Häufigkeit von Überweisungen 
 \newline
Verweis zum Zahnarzt? Alle ausgewertet.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u6$F13)

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U6'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak
Verweis zum Zahnarzt? Nur ausgewertet, wenn Auffälligkeiten an Zähnen oder der Schleimhaut angegeben wurden.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u6$F13[u6$F12==1])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U6'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak
Verweis zum Zahnarzt? Nur ausgewertet, wenn **keine** Auffälligkeiten an Zähnen oder der Schleimhaut angegeben wurden.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl <- tbl_janein(u6$F13[u6$F12==2])

tbl <- ges_janein(tbl)
tbl$x <- rownames_janein
names(tbl) <- c('Antwort', rep(c('n', '%'), 2))

tbl %>% kbl2(booktabs = T, centering=F, align=align_string(tbl)) %>%
  add_header_above(c('', 'U6'=2, 'Ges.'=2)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


## 4.5	Untersuchung auf Sprach- und Sprechstörungen (U5–U9)

### 4.5.2	Hinweise auf Sprach- und Sprechstörungen

Hinweise in der Anamnese?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F12)
tbl2 <- tbl_janein(u6$F14)
tbl3 <- tbl_janein(u7$F12)
tbl4 <- tbl_janein(u7a$F21)
tbl5 <- tbl_janein(u8$F32)
tbl6 <- tbl_janein(u9$F21)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Hinweise in der Orientierenden Beurteilung der Entwicklung?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F13)
tbl2 <- tbl_janein(u6$F15)
tbl3 <- tbl_janein(u7$F13)
tbl4 <- tbl_janein(u7a$F22)
tbl5 <- tbl_janein(u8$F33)
tbl6 <- tbl_janein(u9$F22)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Hinweise entweder in der Anamnese oder in der Orientierenden Beurteilung der Entwicklung?
```{r echo=FALSE, message=FALSE, warning=FALSE}
u5 <- u5 %>% mutate(either=case_when(F12==1 | F13 == 1 ~ 1,   TRUE ~ 2))
tbl1 <- tbl_janein(u5$either)
u6 <- u6 %>% mutate(either=case_when(F14==1 | F15 == 1 ~ 1,   TRUE ~ 2))
tbl2 <- tbl_janein(u6$either)
u7 <- u7 %>% mutate(either=case_when(F12==1 | F13 == 1 ~ 1,   TRUE ~ 2))
tbl3 <- tbl_janein(u7$either)
u7a <- u7a %>% mutate(either=case_when(F21==1 | F22 == 1 ~ 1,   TRUE ~ 2))
tbl4 <- tbl_janein(u7a$either)
u8 <- u8 %>% mutate(either=case_when(F32==1 | F33 == 1 ~ 1,   TRUE ~ 2))
tbl5 <- tbl_janein(u8$either)
u9 <- u9 %>% mutate(either=case_when(F21==1 | F22 == 1 ~ 1,   TRUE ~ 2))
tbl6 <- tbl_janein(u9$either)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


  \linebreak

Überlappung: Hinweise sowohl in der Anamnese als auch in der Orientierenden Beurteilung der Entwicklung?
```{r echo=FALSE, message=FALSE, warning=FALSE}
u5 <- u5 %>% mutate(either=case_when(F12==1 & F13 == 1 ~ 1,   TRUE ~ 2))
tbl1 <- tbl_janein(u5$either)
u6 <- u6 %>% mutate(either=case_when(F14==1 & F15 == 1 ~ 1,   TRUE ~ 2))
tbl2 <- tbl_janein(u6$either)
u7 <- u7 %>% mutate(either=case_when(F12==1 & F13 == 1 ~ 1,   TRUE ~ 2))
tbl3 <- tbl_janein(u7$either)
u7a <- u7a %>% mutate(either=case_when(F21==1 & F22 == 1 ~ 1,   TRUE ~ 2))
tbl4 <- tbl_janein(u7a$either)
u8 <- u8 %>% mutate(either=case_when(F32==1 & F33 == 1 ~ 1,   TRUE ~ 2))
tbl5 <- tbl_janein(u8$either)
u9 <- u9 %>% mutate(either=case_when(F21==1 & F22 == 1 ~ 1,   TRUE ~ 2))
tbl6 <- tbl_janein(u9$either)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Überlappung: Bei Hinweisen in der Orientierenden Beurteilung der Entwicklung auch Hinweise in der Anamnese festgestellt?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F12[u5$F13==1])
tbl2 <- tbl_janein(u6$F14[u6$F15==1])
tbl3 <- tbl_janein(u7$F12[u7$F13==1])
tbl4 <- tbl_janein(u7a$F21[u7a$F22==1])
tbl5 <- tbl_janein(u8$F32[u8$F33==1])
tbl6 <- tbl_janein(u9$F21[u9$F22==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Überlappung: Bei Hinweisen in der Anamnese Hinweise auch in der Orientierenden Beurteilung der Entwicklung?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F13[u5$F12==1])
tbl2 <- tbl_janein(u6$F15[u6$F14==1])
tbl3 <- tbl_janein(u7$F13[u7$F12==1])
tbl4 <- tbl_janein(u7a$F22[u7a$F21==1])
tbl5 <- tbl_janein(u8$F33[u8$F32==1])
tbl6 <- tbl_janein(u9$F22[u9$F21==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


### 4.5.5	Durchführung von Sprachberatung

Besteht erweiterter Beratungsbedarf zum Thema Sprache? Alle ausgewertet.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F14)
tbl2 <- tbl_janein(u6$F16)
tbl3 <- tbl_janein(u7$F14)
tbl4 <- tbl_janein(u7a$F23)
tbl5 <- tbl_janein(u8$F34)
tbl6 <- tbl_janein(u9$F23)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Besteht erweiterter Beratungsbedarf zum Thema Sprache? Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen entweder in der Anamnese oder in der Orientierenden Beurteilung der Entwicklung festgestellt wurden.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F14[u5$F12==1 | u5$F13==1])
tbl2 <- tbl_janein(u6$F16[u6$F14==1 | u6$F15==1])
tbl3 <- tbl_janein(u7$F14[u7$F12==1 | u7$F13==1])
tbl4 <- tbl_janein(u7a$F23[u7a$F21==1 | u7a$F22==1])
tbl5 <- tbl_janein(u8$F34[u8$F32==1 | u8$F33==1])
tbl6 <- tbl_janein(u9$F23[u9$F21==1 | u9$F22==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Besteht erweiterter Beratungsbedarf zum Thema Sprache? Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen in der Anamnese festgestellt wurden. 
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F14[u5$F12==1])
tbl2 <- tbl_janein(u6$F16[u6$F14==1])
tbl3 <- tbl_janein(u7$F14[u7$F12==1])
tbl4 <- tbl_janein(u7a$F23[u7a$F21==1])
tbl5 <- tbl_janein(u8$F34[u8$F32==1])
tbl6 <- tbl_janein(u9$F23[u9$F21==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Besteht erweiterter Beratungsbedarf zum Thema Sprache? Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen in der Orientierenden Beurteilung der Entwicklung festgestellt wurden. 
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F14[u5$F13==1])
tbl2 <- tbl_janein(u6$F16[u6$F15==1])
tbl3 <- tbl_janein(u7$F14[u7$F13==1])
tbl4 <- tbl_janein(u7a$F23[u7a$F22==1])
tbl5 <- tbl_janein(u8$F34[u8$F33==1])
tbl6 <- tbl_janein(u9$F23[u9$F22==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak

Besteht erweiterter Beratungsbedarf zum Thema Sprache? Nur ausgewertet, wenn **keine** Hinweise auf Sprach- und Sprechstörungen weder in der Anamnese noch in der Orientierenden Beurteilung der Entwicklung festgestellt wurden. 
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F14[u5$F12==2 & u5$F13==2])
tbl2 <- tbl_janein(u6$F16[u6$F14==2 & u6$F15==2])
tbl3 <- tbl_janein(u7$F14[u7$F12==2 & u7$F13==2])
tbl4 <- tbl_janein(u7a$F23[u7a$F21==2 & u7a$F22==2])
tbl5 <- tbl_janein(u8$F34[u8$F32==2 & u8$F33==2])
tbl6 <- tbl_janein(u9$F23[u9$F21==2 & u9$F22==2])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak


### 4.5.6	Diagnostische und therapeutische Maßnahmen

Diagnostische/therapeutische Maßnahmen eingeleitet? Alle ausgewertet.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F15)
tbl2 <- tbl_janein(u6$F17)
tbl3 <- tbl_janein(u7$F15)
tbl4 <- tbl_janein(u7a$F24)
tbl5 <- tbl_janein(u8$F35)
tbl6 <- tbl_janein(u9$F24)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


  \linebreak

Bei Hinweisen auf Sprach- und Sprechstörungen: Diagnostische/therapeutische Maßnahmen eingeleitet? Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen entweder in der Anamnese oder in der Orientierenden Beurteilung der Entwicklung festgestellt wurden.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F15[u5$F12==1 | u5$F13==1])
tbl2 <- tbl_janein(u6$F17[u6$F14==1 | u6$F15==1])
tbl3 <- tbl_janein(u7$F15[u7$F12==1 | u7$F13==1])
tbl4 <- tbl_janein(u7a$F24[u7a$F21==1 | u7a$F22==1])
tbl5 <- tbl_janein(u8$F35[u8$F32==1 | u8$F33==1])
tbl6 <- tbl_janein(u9$F24[u9$F21==1 | u9$F22==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


  \linebreak
Bei Hinweisen auf Sprach- und Sprechstörungen: Diagnostische/therapeutische Maßnahmen eingeleitet? Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen in der Anamnese festgestellt wurden.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F15[u5$F12==1])
tbl2 <- tbl_janein(u6$F17[u6$F14==1])
tbl3 <- tbl_janein(u7$F15[u7$F12==1])
tbl4 <- tbl_janein(u7a$F24[u7a$F21==1])
tbl5 <- tbl_janein(u8$F35[u8$F32==1])
tbl6 <- tbl_janein(u9$F24[u9$F21==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak
Bei Hinweisen auf Sprach- und Sprechstörungen: Diagnostische/therapeutische Maßnahmen eingeleitet? Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen in der Orientierenden Beurteilung der Entwicklung festgestellt wurden.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F15[u5$F13==1])
tbl2 <- tbl_janein(u6$F17[u6$F15==1])
tbl3 <- tbl_janein(u7$F15[u7$F13==1])
tbl4 <- tbl_janein(u7a$F24[u7a$F22==1])
tbl5 <- tbl_janein(u8$F35[u8$F33==1])
tbl6 <- tbl_janein(u9$F24[u9$F22==1])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


  \linebreak
 Diagnostische/therapeutische Maßnahmen eingeleitet? Nur ausgewertet, wenn **keine** Hinweise auf Sprach- und Sprechstörungen weder in der Anamnese noch in der Orientierenden Beurteilung der Entwicklung festgestellt wurden.
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u5$F15[u5$F12==2 & u5$F13==2])
tbl2 <- tbl_janein(u6$F17[u6$F14==2 & u6$F15==2])
tbl3 <- tbl_janein(u7$F15[u7$F12==2 & u7$F13==2])
tbl4 <- tbl_janein(u7a$F24[u7a$F21==2 & u7a$F22==2])
tbl5 <- tbl_janein(u8$F35[u8$F32==2 & u8$F33==2])
tbl6 <- tbl_janein(u9$F24[u9$F21==2 & u9$F22==2])
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c('ja', 'nein', 'valides n')) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```



  \linebreak

Wenn diagnostische/therapeutische Maßnahmen eingeleitet: Welche? (Mehrfachauswahl möglich). Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen entweder im Rahmen der Anamnese oder im Rahmen der Orientierenden Beurteilung der Entwicklung angegeben und die Frage nach Einleitung von diagnostischen oder therapeutischen Maßnahmen bejaht wurde. 


```{r message=FALSE, warning=FALSE}
## u5
tbl1 <- u5 %>% 
  filter((F12==1 | F13==1) & F15==1) %>%
  select(F16:F20) %>% 
  mutate(across(everything(), ~replace(., . == 99 , 0))) %>% 
  mutate(valid=ifelse(rowSums(. ,na.rm=TRUE )>=1,1,0), 
         missing=ifelse(rowSums(., na.rm=TRUE)==0,1,0), 
         ges=1) %>%
  summarise_all(sum, na.rm = TRUE) %>%
  set_names(linebreak(c(
    "01 Logopädie",
    "02 Pädaudiologie / Phoniatrie",
    "03 Sozialpädiatrisches Zentrum (SPZ)",
    "04 HNO-Arzt",
    "05 Sonstige",
    "97 valides n", 
    "98 fehlend",
    "99 gesamt n"))) %>% 
  rownames_to_column() %>% 
  pivot_longer(-rowname, names_to = "var", values_to= "n") %>%
  #spread(Gruppe_, n) # nur wenn mehrere Gruppen
  select(var, n) %>% 
  mutate(var=str_sub(var, 4L)) %>%
  set_names(c("Maßnahme", "n")) #%>%
  
# Prozentangaben berechnen
tbl1$freq <- (tbl1$n / tbl1$n[tbl1$Maßnahme=="valides n"])*100
# ungültige löschen
tbl1[nrow(tbl1),'freq'] <- NA
tbl1[nrow(tbl1)-1,'freq'] <- NA
tbl1[nrow(tbl1)-2,'freq'] <- NA

## u6
tbl2 <- u6 %>% 
  filter((F14==1 | F15==1) & F17==1) %>%
  select(F18:F22) %>% 
  mutate(across(everything(), ~replace(., . == 99 , 0))) %>% 
  mutate(valid=ifelse(rowSums(. ,na.rm=TRUE )>=1,1,0), 
         missing=ifelse(rowSums(., na.rm=TRUE)==0,1,0), 
         ges=1) %>%
  summarise_all(sum, na.rm = TRUE) %>%
  set_names(linebreak(c(
    "01 Logopädie",
    "02 Pädaudiologie / Phoniatrie",
    "03 Sozialpädiatrisches Zentrum (SPZ)",
    "04 HNO-Arzt",
    "05 Sonstige",
    "97 valides n", 
    "98 fehlend",
    "99 gesamt n"))) %>% 
  rownames_to_column() %>% 
  pivot_longer(-rowname, names_to = "var", values_to= "n") %>%
  #spread(Gruppe_, n) # nur wenn mehrere Gruppen
  select(var, n) %>% 
  mutate(var=str_sub(var, 4L)) %>%
  set_names(c("Maßnahme", "n")) #%>%
  
# Prozentangaben berechnen
tbl2$freq <- (tbl2$n / tbl2$n[tbl2$Maßnahme=="valides n"])*100
# ungültige löschen
tbl2[nrow(tbl2),'freq'] <- NA
tbl2[nrow(tbl2)-1,'freq'] <- NA
tbl2[nrow(tbl1)-2,'freq'] <- NA


## u7
tbl3 <- u7 %>% 
  filter((F12==1 | F13==1) & F15==1) %>%
  select(F16:F20) %>% 
  mutate(across(everything(), ~replace(., . == 99 , 0))) %>% 
  mutate(valid=ifelse(rowSums(. ,na.rm=TRUE )>=1,1,0), 
         missing=ifelse(rowSums(., na.rm=TRUE)==0,1,0), 
         ges=1) %>%
  summarise_all(sum, na.rm = TRUE) %>%
  set_names(linebreak(c(
    "01 Logopädie",
    "02 Pädaudiologie / Phoniatrie",
    "03 Sozialpädiatrisches Zentrum (SPZ)",
    "04 HNO-Arzt",
    "05 Sonstige",
    "97 valides n", 
    "98 fehlend",
    "99 gesamt n"))) %>% 
  rownames_to_column() %>% 
  pivot_longer(-rowname, names_to = "var", values_to= "n") %>%
  #spread(Gruppe_, n) # nur wenn mehrere Gruppen
  select(var, n) %>% 
  mutate(var=str_sub(var, 4L)) %>%
  set_names(c("Maßnahme", "n")) #%>%
  
# Prozentangaben berechnen
tbl3$freq <- (tbl3$n / tbl3$n[tbl3$Maßnahme=="valides n"])*100
# ungültige löschen
tbl3[nrow(tbl3),'freq'] <- NA
tbl3[nrow(tbl3)-1,'freq'] <- NA
tbl3[nrow(tbl1)-2,'freq'] <- NA


## u7a
tbl4 <- u7a %>% 
  filter((F21==1 | F22==1) & F24==1) %>%
  select(F25:F29) %>% 
  mutate(across(everything(), ~replace(., . == 99 , 0))) %>% 
  mutate(valid=ifelse(rowSums(. ,na.rm=TRUE )>=1,1,0), 
         missing=ifelse(rowSums(., na.rm=TRUE)==0,1,0), 
         ges=1) %>%
  summarise_all(sum, na.rm = TRUE) %>%
  set_names(linebreak(c(
    "01 Logopädie",
    "02 Pädaudiologie / Phoniatrie",
    "03 Sozialpädiatrisches Zentrum (SPZ)",
    "04 HNO-Arzt",
    "05 Sonstige",
    "97 valides n", 
    "98 fehlend",
    "99 gesamt n"))) %>% 
  rownames_to_column() %>% 
  pivot_longer(-rowname, names_to = "var", values_to= "n") %>%
  #spread(Gruppe_, n) # nur wenn mehrere Gruppen
  select(var, n) %>% 
  mutate(var=str_sub(var, 4L)) %>%
  set_names(c("Maßnahme", "n")) #%>%
  
# Prozentangaben berechnen
tbl4$freq <- (tbl4$n / tbl4$n[tbl4$Maßnahme=="valides n"])*100
# ungültige löschen
tbl4[nrow(tbl4),'freq'] <- NA
tbl4[nrow(tbl4)-1,'freq'] <- NA
tbl4[nrow(tbl1)-2,'freq'] <- NA


## u8
tbl5 <- u8 %>% 
  filter((F32==1 | F33==1) & F35==1) %>%
  select(F36:F40) %>% 
  mutate(across(everything(), ~replace(., . == 99 , 0))) %>% 
  mutate(valid=ifelse(rowSums(. ,na.rm=TRUE )>=1,1,0), 
         missing=ifelse(rowSums(., na.rm=TRUE)==0,1,0), 
         ges=1) %>%
  summarise_all(sum, na.rm = TRUE) %>%
  set_names(linebreak(c(
    "01 Logopädie",
    "02 Pädaudiologie / Phoniatrie",
    "03 Sozialpädiatrisches Zentrum (SPZ)",
    "04 HNO-Arzt",
    "05 Sonstige",
    "97 valides n", 
    "98 fehlend",
    "99 gesamt n"))) %>% 
  rownames_to_column() %>% 
  pivot_longer(-rowname, names_to = "var", values_to= "n") %>%
  #spread(Gruppe_, n) # nur wenn mehrere Gruppen
  select(var, n) %>% 
  mutate(var=str_sub(var, 4L)) %>%
  set_names(c("Maßnahme", "n")) #%>%
  
# Prozentangaben berechnen
tbl5$freq <- (tbl5$n / tbl5$n[tbl5$Maßnahme=="valides n"])*100
# ungültige löschen
tbl5[nrow(tbl5),'freq'] <- NA
tbl5[nrow(tbl5)-1,'freq'] <- NA
tbl5[nrow(tbl1)-2,'freq'] <- NA

## u9
tbl6 <- u9 %>% 
  filter((F21==1 | F22==1) & F24==1) %>%
  select(F25:F29) %>% 
  mutate(across(everything(), ~replace(., . == 99 , 0))) %>% 
  mutate(valid=ifelse(rowSums(. ,na.rm=TRUE )>=1,1,0), 
         missing=ifelse(rowSums(., na.rm=TRUE)==0,1,0), 
         ges=1) %>%
  summarise_all(sum, na.rm = TRUE) %>%
  set_names(linebreak(c(
    "01 Logopädie",
    "02 Pädaudiologie / Phoniatrie",
    "03 Sozialpädiatrisches Zentrum (SPZ)",
    "04 HNO-Arzt",
    "05 Sonstige",
    "97 valides n", 
    "98 fehlend",
    "99 gesamt n"))) %>% 
  rownames_to_column() %>% 
  pivot_longer(-rowname, names_to = "var", values_to= "n") %>%
  #spread(Gruppe_, n) # nur wenn mehrere Gruppen
  select(var, n) %>% 
  mutate(var=str_sub(var, 4L)) %>%
  set_names(c("Maßnahme", "n")) #%>%
  
# Prozentangaben berechnen
tbl6$freq <- (tbl6$n / tbl6$n[tbl6$Maßnahme=="valides n"])*100
# ungültige löschen
tbl6[nrow(tbl6),'freq'] <- NA
tbl6[nrow(tbl6)-1,'freq'] <- NA
tbl6[nrow(tbl1)-2,'freq'] <- NA


tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3])
names(tbl) <- c('Antwort','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]')

# Gesamt-Spalte einfügen
tbl <- tbl %>% mutate(`Ges. [n]` = rowSums(across(contains('[n]'))))
tbl$`Ges. [%]` <- tbl$`Ges. [n]` / tbl$`Ges. [n]`[6]*100
tbl$`Ges. [%]`[6] <- NA
tbl$`Ges. [%]`[7] <- NA
tbl$`Ges. [%]`[8] <- NA

# der Häufigkeit nach sortieren
tbl1 <- tbl[1:4,]
tbl1 <- tbl1 %>% arrange(desc(`Ges. [n]`))
tbl[1:4,] <- tbl1

# Fallzahlen
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  slice_tail(n=3) %>% 
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left")

# absolute Häufigkeiten
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  slice_head(n=5) %>%
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left")

# relative Häufigkeiten
tmp <- tbl %>% dplyr::select(Antwort, contains('[%]')) %>%
  filter(!Antwort %in% c('valides n', 'fehlend', 'gesamt n')) 

tmp %>%
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left")

openxlsx::write.xlsx(tmp, file='05_Tabellen/Sprache_Massnahmen.xlsx')
```

\pagebreak
Analyse von Freitexteintragungen unter Sonstige. Nur ausgewertet, wenn Hinweise auf Sprach- und Sprechstörungen entweder im Rahmen der Anamnese oder im Rahmen der Orientierenden Beurteilung der Entwicklung angegeben, die Frage nach Einleitung von diagnostischen oder therapeutischen Maßnahmen bejaht und sonstige diagnostische/therapeutische Maßnahmen angekreuzt wurde. 
```{r sprache_freitext, message=FALSE, warning=FALSE}
### Fälle filtern
tbl1 <- u5 %>% filter((F12==1 | F13==1) & F15==1 & F20==1) %>% mutate(id=paste0(id, U, kind)) %>% select(id, U)
tbl2 <- u6 %>% filter((F14==1 | F15==1) & F17==1 & F22==1) %>% mutate(id=paste0(id, U, kind)) %>% select(id, U)
tbl3 <- u7 %>% filter((F12==1 | F13==1) & F15==1 & F20==1) %>% mutate(id=paste0(id, U, kind)) %>% select(id, U)
tbl4 <- u7a %>% filter((F21==1 | F22==1) & F24==1 & F29==1) %>% mutate(id=paste0(id, U, kind)) %>% select(id, U)
tbl5 <- u8 %>% filter((F32==1 | F33==1) & F35==1 & F40==1) %>% mutate(id=paste0(id, U, kind)) %>% select(id, U)
tbl6 <- u9 %>% filter((F21==1 | F22==1) & F24==1 & F29==1) %>% mutate(id=paste0(id, U, kind)) %>% select(id, U)

tbl <- bind_rows(tbl1, tbl2, tbl3, tbl4, tbl5, tbl6)

### MaxQDA-Ergebnis einlesen
txt <- openxlsx::read.xlsx("O:/U4951/Consulimus_Abschlussdatenlieferung/03_Sammelbögen/02_Freitexte/FreitextSprache_Dokumente&Codes_grb.xlsx")

txt <- txt %>% 
  mutate(id = str_extract(Dokument, '^[^,]*')) %>% 
  separate(Codes, "\n", into = c("U", 'Codes1', 'Codes2', 'Codes3'), extra = "merge") %>% 
  select(id, U, Codes1, Codes2, Codes3)

# txt nach Fällen filtern
txt2 <- txt %>% filter(id %in% tbl$id) %>% filter(!is.na(Codes1)) #n = 995 Kinder mit Freitexten

### Tabelle über Anzahl Personen mit Freitexten
tbl %>% 
  mutate(`Freitext angegeben` = ifelse(id %in% txt2$id, 'Ja', 'Nein')) %>% 
  tabyl(`Freitext angegeben`, U) %>% 
  adorn_totals('row', name='Gesamt n') %>% 
  adorn_totals('col', name='Ges.') %>% 
  kbl2(booktabs = T, row.names = F) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>% 
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

### selektierte Fälle auswerten
txt2 <- txt2 %>% 
  pivot_longer(cols = starts_with('Codes'), names_to = 'pos', values_to = 'Code') %>%
  na.omit() %>%  
  group_by(U, Code) %>% 
  tally() %>% 
  pivot_wider(names_from = U, values_from = n) %>% 
  janitor::adorn_totals('col', name = 'Ges.') %>% 
  # Kategroeien umbenennen
  mutate(Code = ifelse(Code == 'Soziale Maßnahmen', 'Nicht-medizinsiche Maßnahmen', Code))

### abs
txt2 %>% 
 mutate(across(U5:`Ges.`, ~replace_na(., 0))) %>% 
 kbl2(booktabs = T, digits=2, row.names = F, 
      col.names= c('Antwort', 'U5 [n]', 'U6 [n]','U7 [n]','U7a [n]','U8 [n]','U9 [n]', 'Ges. [n]')) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left")
```

Prozentangaben beziehen sich auf die Anzahl von U-Untersuchungen mit Angabe von Freitexten zu sonstigen diagnostischen / therapeutischen Maßnahmen.
```{r sprache_freitext2, message=FALSE, warning=FALSE}
### rel
txt2 %>% 
  mutate(across(U5:`Ges.`, ~replace_na(., 0))) %>% 
  mutate(U5 = U5 / 6 * 100, 
         U6 = U6 / 11 * 100,
         U7 = U7 / 222 * 100,
         U7a = U7a / 222 * 100,
         U8 = U8 / 278 * 100,
         U9 = U9 / 256 *100,
         `Ges.` = `Ges.` / 995 *100) %>% 
  kbl2(booktabs = T, digits=1, row.names = F, 
      col.names= c('Antwort', 'U5 [%]', 'U6 [%]','U7 [%]','U7a [%]','U8 [%]','U9 [%]', 'Ges. [%]')) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left")
```


Prozentangaben beziehen sich auf die Anzahl von U-Untersuchungen mit irgendwelchen diagnostischen / therapeutischen Maßnahmen
(wenn Hinweise auf Sprach- und Sprechstörungen entweder im Rahmen der Anamnese oder im Rahmen der Orientierenden Beurteilung der Entwicklung angegeben wurden).
```{r sprache_freitext3, message=FALSE, warning=FALSE}
### rel
txt2 %>% 
  mutate(across(U5:`Ges.`, ~replace_na(., 0))) %>% 
  mutate(U5 = U5 / 26 * 100, 
         U6 = U6 / 32 * 100,
         U7 = U7 / 436 * 100,
         U7a = U7a / 618 * 100,
         U8 = U8 / 966 * 100,
         U9 = U9 / 1208 *100,
         `Ges.` = `Ges.` / 3286 *100) %>% 
  kbl2(booktabs = T, digits=1, row.names = F, 
      col.names= c('Antwort', 'U5 [%]', 'U6 [%]','U7 [%]','U7a [%]','U8 [%]','U9 [%]', 'Ges. [%]')) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left")
```


```{r echo=FALSE, message=FALSE, warning=FALSE}
### Prüfung, ob nicht-medizinische Maßnahmen & Kontrolle disjunkt
# txt2 <- txt %>% 
#   filter(id %in% tbl$id) %>% 
#   filter(!is.na(Codes1)) %>% 
#   pivot_longer(cols = starts_with('Codes'), names_to = 'pos', values_to = 'Code') %>%
#   na.omit() %>%  
#   group_by(id) %>% 
#   mutate(abwarten = ifelse(any(Code == 'Beobachtung/Kontrolle'),1,0), 
#          sozial = ifelse(any(Code == 'Soziale Maßnahmen'),1,0), 
#          beides = ifelse((abwarten + sozial) == 2, 1,0)) %>% 
#   ungroup() %>%
#   distinct(id, .keep_all = T) %>% 
#   mutate_at(vars(abwarten, sozial, beides), as.factor) %>% 
#   group_by(abwarten, sozial, beides) %>% 
#   tally()

#kein Output, bisher nur angeguckt
```


\pagebreak
## 4.7	Information und Beratung der Eltern (U2–U9)

### 4.7.2	Feststellung von erweitertem Beratungsbedarf (U2–U9)
Wie viele Beratungsthemen mit erweitertem Beratungsbedarf wurden im Abschnitt *Beratung* angekreuzt?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_kont(u2$F5)
tbl2 <- tbl_kont(u3$F5)
tbl3 <- tbl_kont(u4$F5)
tbl4 <- tbl_kont(u5$F5)
tbl5 <- tbl_kont(u6$F5)
tbl6 <- tbl_kont(u7$F5)
tbl7 <- tbl_kont(u7a$F5)
tbl8 <- tbl_kont(u8$F5)
tbl9 <- tbl_kont(u9$F5)
tbl <- bind_rows(tbl1, tbl2, tbl3, tbl4, tbl5, tbl6, tbl7, tbl8, tbl9)

tbl$U <- c('U2', 'U3', 'U4', 'U5', 'U6', 'U7', 'U7a', 'U8', 'U9')
tbl <- select(tbl, U, everything())

tbl %>% 
  kbl2(booktabs = T, digits=2, row.names = F) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") # Zwingt die Tabelle an die Position im Text
```
  \linebreak

Anteil der Kinder mit mindestens einem erweiterten Beratungsbedarf
```{r echo=FALSE, message=FALSE, warning=FALSE}
u2$F5_jn <- ifelse(u2$F5==0, 2, 1)
u3$F5_jn <- ifelse(u3$F5==0, 2, 1)
u4$F5_jn <- ifelse(u4$F5==0, 2, 1)
u5$F5_jn <- ifelse(u5$F5==0, 2, 1)
u6$F5_jn <- ifelse(u6$F5==0, 2, 1)
u7$F5_jn <- ifelse(u7$F5==0, 2, 1)
u7a$F5_jn <- ifelse(u7a$F5==0, 2, 1)
u8$F5_jn <- ifelse(u8$F5==0, 2, 1)
u9$F5_jn <- ifelse(u9$F5==0, 2, 1)

tbl1 <- tbl_janein(u2$F5_jn)
tbl2 <- tbl_janein(u3$F5_jn)
tbl3 <- tbl_janein(u4$F5_jn)
tbl4 <- tbl_janein(u5$F5_jn)
tbl5 <- tbl_janein(u6$F5_jn)
tbl6 <- tbl_janein(u7$F5_jn)
tbl7 <- tbl_janein(u7a$F5_jn)
tbl8 <- tbl_janein(u8$F5_jn)
tbl9 <- tbl_janein(u9$F5_jn)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

#tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')
tbl$Antwort <- c(">= 1 Themen", "keine Themen", "valides n", "fehlend", "gesamt n" )

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c(">= 1 Themen", "keine Themen", "valides n")) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak
Anzahl genannter Beratungsthemen: 
```{r echo=FALSE, message=FALSE, warning=FALSE}
u2$F5_cat <- ifelse(u2$F5 >= 4, '4+', u2$F5)
u3$F5_cat <- ifelse(u3$F5 >= 4, '4+', u3$F5)
u4$F5_cat <- ifelse(u4$F5 >= 4, '4+', u4$F5)
u5$F5_cat <- ifelse(u5$F5 >= 4, '4+', u5$F5)
u6$F5_cat <- ifelse(u6$F5 >= 4, '4+', u6$F5)
u7$F5_cat <- ifelse(u7$F5 >= 4, '4+', u7$F5)
u7a$F5_cat <- ifelse(u7a$F5 >= 4, '4+', u7a$F5)
u8$F5_cat <- ifelse(u8$F5 >= 4, '4+', u8$F5)
u9$F5_cat <- ifelse(u9$F5 >= 4, '4+', u9$F5)

tbl1 <- u2 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl2 <- u3 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl3 <- u4 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl4 <- u5 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl5 <- u6 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl6 <- u7 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl7 <- u7a %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl8 <- u8 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl9 <- u9 %>% filter(F5 != 0) %>% select(F5_cat) %>% tbl_mehrf()
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]',
                'U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]')

tbl <- ges_mehrf(tbl)


tbl$Antwort <- as.character(tbl$Antwort)
tbl$Antwort[1:(nrow(tbl)-3)] <- paste(tbl$Antwort[1:(nrow(tbl)-3)], 'Themen', sep=' ')
tbl$Antwort[(nrow(tbl)-2):nrow(tbl)] <- c("valides n", "fehlend", "gesamt n")

tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") %>% # Zwingt die Tabelle an die Position im Text
  row_spec(4, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(!Antwort %in% c("fehlend", "gesamt n")) %>% 
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") %>% # Zwingt die Tabelle an die Position im Text
  row_spec(4, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


# 5. Ergebnisse der Evaluation von Anlage 1 (Untersuchungsheft für Kinder) der Kinder-Richtlinie

## 5.2	Vorliegen des Gelben Heftes bei U-Untersuchungen

Lag das Gelbe Heft bei der Untersuchung vor?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F1)
tbl2 <- tbl_janein(u3$F1)
tbl3 <- tbl_janein(u4$F1)
tbl4 <- tbl_janein(u5$F1)
tbl5 <- tbl_janein(u6$F1)
tbl6 <- tbl_janein(u7$F1)
tbl7 <- tbl_janein(u7a$F1)
tbl8 <- tbl_janein(u8$F1)
tbl9 <- tbl_janein(u9$F1)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')


# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c("ja", "nein", "valides n")) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```




## 5.3	Häufigkeit der Dokumentation im Gelben Heft

### 5.3.1	Dokumentation von Auffälligkeiten 

#### 5.3.1.1 Auffälligkeiten im Abschnitt *Untersuchung* insgesamt 

 \newline
Wie viele Auffälligkeiten wurden im Abschnitt *Untersuchung* angekreuzt?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_kont(u2$F4)
tbl2 <- tbl_kont(u3$F4)
tbl3 <- tbl_kont(u4$F4)
tbl4 <- tbl_kont(u5$F4)
tbl5 <- tbl_kont(u6$F4)
tbl6 <- tbl_kont(u7$F4)
tbl7 <- tbl_kont(u7a$F4)
tbl8 <- tbl_kont(u8$F4)
tbl9 <- tbl_kont(u9$F4)
tbl <- bind_rows(tbl1, tbl2, tbl3, tbl4, tbl5, tbl6, tbl7, tbl8, tbl9)

tbl$U <- c('U2', 'U3', 'U4', 'U5', 'U6', 'U7', 'U7a', 'U8', 'U9')
tbl <- select(tbl, U, everything())

tbl %>% 
  kbl2(booktabs = T, digits=2, row.names = F) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") # Zwingt die Tabelle an die Position im Text
```

  \linebreak
Anzahl genannter Auffälligkeiten im Abschnitt *Untersuchung*: 
```{r echo=FALSE, message=FALSE, warning=FALSE}
u2$F4_cat <- ifelse(u2$F4 >= 3, '3+', u2$F4)
u3$F4_cat <- ifelse(u3$F4 >= 3, '3+', u3$F4)
u4$F4_cat <- ifelse(u4$F4 >= 3, '3+', u4$F4)
u5$F4_cat <- ifelse(u5$F4 >= 3, '3+', u5$F4)
u6$F4_cat <- ifelse(u6$F4 >= 3, '3+', u6$F4)
u7$F4_cat <- ifelse(u7$F4 >= 3, '3+', u7$F4)
u7a$F4_cat <- ifelse(u7a$F4 >= 3, '3+', u7a$F4)
u8$F4_cat <- ifelse(u8$F4 >= 3, '3+', u8$F4)
u9$F4_cat <- ifelse(u9$F4 >= 3, '3+', u9$F4)

tbl1 <- tbl_mehrf(u2$F4_cat)
tbl2 <- tbl_mehrf(u3$F4_cat)
tbl3 <- tbl_mehrf(u4$F4_cat)
tbl4 <- tbl_mehrf(u5$F4_cat)
tbl5 <- tbl_mehrf(u6$F4_cat)
tbl6 <- tbl_mehrf(u7$F4_cat)
tbl7 <- tbl_mehrf(u7a$F4_cat)
tbl8 <- tbl_mehrf(u8$F4_cat)
tbl9 <- tbl_mehrf(u9$F4_cat)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]',
                'U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]')

tbl <- ges_mehrf(tbl)


tbl$Antwort <- as.character(tbl$Antwort)
tbl$Antwort[1:(nrow(tbl)-3)] <- paste(tbl$Antwort[1:(nrow(tbl)-3)], 'Auffälligkeiten', sep=' ')
tbl$Antwort[(nrow(tbl)-2):nrow(tbl)] <- c("valides n", "fehlend", "gesamt n")

tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") %>% # Zwingt die Tabelle an die Position im Text
  row_spec(4, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(!Antwort %in% c("fehlend", "gesamt n")) %>% 
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") %>% # Zwingt die Tabelle an die Position im Text
  row_spec(4, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```



#### 5.3.1.3	Auffälligkeiten im Abschnitt *Ergebnisse*

 \newline
Eintrag bei „Auffälligkeiten zur Beobachtung“ im Abschnitt *Ergebnisse*?
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F6)
tbl2 <- tbl_janein(u3$F6)
tbl3 <- tbl_janein(u4$F6)
tbl4 <- tbl_janein(u5$F6)
tbl5 <- tbl_janein(u6$F6)
tbl6 <- tbl_janein(u7$F6)
tbl7 <- tbl_janein(u7a$F6)
tbl8 <- tbl_janein(u8$F6)
tbl9 <- tbl_janein(u9$F6)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')


# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c("ja", "nein", "valides n")) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```


### 5.3.2	Dokumentation von erweitertem Beratungsbedarf 

#### 5.3.2.1	Erweiterter Beratungsbedarf insgesamt
 \newline
Wie viele Beratungsthemen mit erweitertem Beratungsbedarf wurden im Abschnitt *Beratung* angekreuzt?
```{r echo=FALSE, message=FALSE, warning=FALSE}

```

Die Frage ist bereits im Abschnitt 3.7.2 ausgewertet


## 5.5	Aufwand für die Dokumentation 

Anzahl genannter Zeitaufwände (eingeteilt in Kategorien)
```{r echo=FALSE, message=FALSE, warning=FALSE}
u2 <- u2 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u3 <- u3 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u4 <- u4 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u5 <- u5 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u6 <- u6 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u7 <- u7 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u7a <- u7a %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u8 <- u8 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))
u9 <- u9 %>% mutate(F8_cat=case_when(F8 == 0 ~ as.character(NA), F8 < 1 ~ '<1', F8 < 2 ~ '>=1 & <2', 
                                     F8 < 5 ~ '>=2 & <5', F8 < 10 ~ '>=5 & <10', F8 >= 10 ~ '>=10'))


tbl1 <- tbl_mehrf(u2$F8_cat)
tbl2 <- tbl_mehrf(u3$F8_cat)
tbl3 <- tbl_mehrf(u4$F8_cat)
tbl4 <- tbl_mehrf(u5$F8_cat)
tbl5 <- tbl_mehrf(u6$F8_cat)
tbl6 <- tbl_mehrf(u7$F8_cat)
tbl7 <- tbl_mehrf(u7a$F8_cat)
tbl8 <- tbl_mehrf(u8$F8_cat)
tbl9 <- tbl_mehrf(u9$F8_cat)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]',
                'U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 
                 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]')

tbl <- ges_mehrf(tbl)

tbl$Antwort <- as.character(tbl$Antwort)
tbl$Antwort[1:(nrow(tbl)-3)] <- paste(tbl$Antwort[1:(nrow(tbl)-3)], 'Min.', sep=' ')
tbl$Antwort[(nrow(tbl)-2):nrow(tbl)] <- c("valides n", "fehlend", "gesamt n")

tbl <- tbl %>% arrange(match(Antwort, c('0 Min.', '<1 Min.', '>=1 & <2 Min.', '>=2 & <5 Min.', 
                                        '>=5 & <10 Min.', '>=10 Min.')))

tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") %>% # Zwingt die Tabelle an die Position im Text
  row_spec(5, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(!Antwort %in% c("fehlend", "gesamt n")) %>% 
  kbl2(booktabs = T, digits=1) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") %>% # Zwingt die Tabelle an die Position im Text
  row_spec(5, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```

  \linebreak
Zeitaufwand für Eintragungen in das Gelbe Heft (Schätzwert in Minuten, nur valide Angaben inkludiert)
```{r echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_kont(u2$F8[u2$F8!=0])
tbl2 <- tbl_kont(u3$F8[u3$F8!=0])
tbl3 <- tbl_kont(u4$F8[u4$F8!=0])
tbl4 <- tbl_kont(u5$F8[u5$F8!=0])
tbl5 <- tbl_kont(u6$F8[u6$F8!=0])
tbl6 <- tbl_kont(u7$F8[u7$F8!=0])
tbl7 <- tbl_kont(u7a$F8[u7a$F8!=0])
tbl8 <- tbl_kont(u8$F8[u8$F8!=0])
tbl9 <- tbl_kont(u9$F8[u9$F8!=0])
tbl <- bind_rows(tbl1, tbl2, tbl3, tbl4, tbl5, tbl6, tbl7, tbl8, tbl9)

tbl$U <- c('U2', 'U3', 'U4', 'U5', 'U6', 'U7', 'U7a', 'U8', 'U9')
tbl <- select(tbl, U, everything())

tbl %>%
  kbl2(booktabs = T, digits=2, row.names = F) %>% # Für den Tabellenstil
  kable_styling(latex_options = "HOLD_position", position = "left") # Zwingt die Tabelle an die Position im Text
```

  \linebreak

## 5.6	Nutzung der Teilnahmekarte

### 5.6.1	Ausfüllen der Teilnahmekarte

Teilnahmekarte ausgefüllt?
```{r teilnahmekarte, echo=FALSE, message=FALSE, warning=FALSE}
tbl1 <- tbl_janein(u2$F7)
tbl2 <- tbl_janein(u3$F7)
tbl3 <- tbl_janein(u4$F7)
tbl4 <- tbl_janein(u5$F7)
tbl5 <- tbl_janein(u6$F7)
tbl6 <- tbl_janein(u7$F7)
tbl7 <- tbl_janein(u7a$F7)
tbl8 <- tbl_janein(u8$F7)
tbl9 <- tbl_janein(u9$F7)
tbl <- bind_cols(tbl1, tbl2[2:3], tbl3[2:3], tbl4[2:3], tbl5[2:3], tbl6[2:3], tbl7[2:3], tbl8[2:3], tbl9[2:3])

tbl$x <- rownames_janein
tbl <- ges_janein(tbl)
names(tbl) <- c('Antwort','U2 [n]','U2 [%]','U3 [n]','U3 [%]','U4 [n]','U4 [%]','U5 [n]','U5 [%]','U6 [n]','U6 [%]','U7 [n]','U7 [%]', 'U7a [n]','U7a [%]','U8 [n]','U8 [%]','U9 [n]','U9 [%]', 'Ges. [n]', 'Ges. [%]')

# absolute Tabelle
tbl %>% dplyr::select(Antwort, contains('[n]')) %>%
  kbl2(booktabs = T, digits=1) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)

# relative Tabelle
tbl %>% dplyr::select(Antwort, contains('[%]')) %>% 
  filter(Antwort %in% c("ja", "nein", "valides n")) %>% 
  kbl2(booktabs = T, digits=1, align=align_string(tbl)) %>% 
  kable_styling(latex_options = "HOLD_position", position = "left") %>%
  row_spec(2, extra_css = "border-bottom: 1px solid", hline_after=TRUE)
```
